pyproject.toml
[tool.poetry]
name = "meltano-file-parser"
version = "0.0.1"
description = ""
authors = ["Edgar Ramirez-Mondragon <edgarrm358@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.9.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = "models"
